Update README.md
https://internship.suvenconsultants.com for providing awesome problem statements and giving many of us a Coding Internship Experience. @www.suvenconsultants.com"

Conclusion Global warming is deteriorating the climate and is affecting various parameters of the environment. Hence from this analysis, we infer that there is either a sharp rise in temperatures or sharp falls over the 10 yrs. Hence we can conclude that Global Warming has caused a major difference and unreliability in temperature predictability also taking humidity into consideration we can say that it has almost remained the same throughout the past years.

WRITTEN BY
Rajeskumar K